backend/requirements.txt
flask-socketio
numpy
pandas
matplotlib
scipy
aiohttp
requests
python-dotenv
quart
hypercorn
python-socketio
python-engineio
# Async web framework
quart
quart-cors
python-socketio

# Async I/O and networking
aiohttp
asyncpg

# Environment and logging
python-dotenv

# Data processing and numerical
numpy

